Add favorite foods and restaurants backend + some basic evalution based on those

+++ app/src/main/java/soft/brunhilda/org/dailymenupicker/entity/EvaluatableEntity.kt
@@ -1,4 +1,4 @@
 package soft.brunhilda.org.dailymenupicker.entity
 
-abstract class EvaluatableEntity(var preferenceEvaluation: Int = 2) {
+abstract class EvaluatableEntity(var preferenceEvaluation: Double = 2.0) {
 }
